authentication_jwt/myproject/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework_simplejwt
black